requirements.txt
aiobotocore[boto3]
fastapi
uvicorn
qdrant-client
sentence-transformers
pillow
tqdm
